subproject_database_manager/requirements.txt
langgraph
pinecone-client
openai
python-dotenv
telethon
